new test images --> not working on rotated docs

+++ scanner.py
@@ -23,42 +23,38 @@ def find_document_corners(img):
     blurred_img = cv.GaussianBlur(img, ksize=(5, 5), sigmaX=0)
 
     # 2. Find all contours
-    print("Edge Detection")
-    edge_img = cv.Canny(blurred_img, 75, 200)
+    # print("Edge Detection")
+    edge_img = cv.Canny(blurred_img, 160, 210)
 
-    print("Contour Detection")
+    # print("Contour Detection")
     # use cv.CHAIN_APPROX_SIMPLE --> best case only 4 points needed for perfect rectangular document
     # ToDo: understand retrieval-mode: cv.RETR_EXTERNAL vs RETR_TREE vs RETR_LIST?!
-    contours, _hierarchy = cv.findContours(edge_img, mode=cv.RETR_LIST, method=cv.CHAIN_APPROX_SIMPLE)
+    contours, _hierarchy = cv.findContours(edge_img, mode=cv.RETR_EXTERNAL, method=cv.CHAIN_APPROX_SIMPLE)
 
-    # BUG: if contour consists of many points the area seems to be unreasonably big --> wrong contour selected
-    # ToDo: document contour may consist of more than 4 points --> non-maximum supression?
-    # 3. Find largest contour
-
-    # approximating the contour with (hopefully) only 4 points
-    # epsilon 0.01 * contour_perimeter seems to work empirically well for us
+    contour_img = img.copy()
+    # 3. Find largest contour which can be approximated using only four points
 
     largest_contour = None
     largest_area = -1
     for contour in contours:
         perimeter = cv.arcLength(contour, closed=True)
-        low_poly_contour = cv.approxPolyDP(contour, 0.01 * perimeter, closed=True)
+        low_poly_contour = cv.approxPolyDP(contour, 0.02 * perimeter, closed=True)
 
         if len(low_poly_contour) == 4:
-            area = cv.contourArea(low_poly_contour)
-            if area > largest_area:
+            (x, y, w, h) = cv.boundingRect(contour)
+            # cv.rectangle(contour_img, (x, y), (x + w, y + h), (255, 0, 0), 2)
+            area = w * h
+            if area > largest_area and area > 10000:
                 largest_area = area
                 largest_contour = low_poly_contour
 
     if largest_contour is None:
-        print("unable to find a countour that can be approximated using 4 corner points!")
-        sys.exit(-1)
+        return None
 
     # 4. extract four corner points of largest contour
     # (assuming largest contour is our document --> if it fails: Hough Transform?)
 
     if DEBUG:
-        contour_img = img.copy()
         cv.drawContours(contour_img, contours, -1, (0, 0, 255), 1)
         cv.drawContours(contour_img, largest_contour, -1, (0, 255, 0), 5)
 
@@ -68,8 +64,8 @@ def find_document_corners(img):
         cv.imshow("Edges", edge_img)
         cv.imshow("Contours", contour_img)
 
-        cv.resizeWindow("Edges", int(1080 * 0.75), 1080)
-        cv.resizeWindow("Contours", int(1080 * 0.75), 1080)
+        cv.resizeWindow("Edges", 707, 1000)
+        cv.resizeWindow("Contours", 707, 1000)
 
         cv.waitKey(0)
         cv.destroyAllWindows()
@@ -81,7 +77,7 @@ def find_document_corners(img):
 
 
 def perspective_transform(img, doc_corners):
-    print("Persective Transformation")
+    # print("Persective Transformation")
     # ToDo: take rotation into account
 
     # Final Image dimensions (keep aspect ratio of DIN A4)!
@@ -89,14 +85,12 @@ def perspective_transform(img, doc_corners):
     HEIGHT = 297 * 5
 
     src = np.array([doc_corners[1][0], doc_corners[2][0], doc_corners[0][0], doc_corners[3][0]], dtype="float32")
-    # print(src)
     dst = np.array([[0, 0], [0, HEIGHT - 1], [WIDTH - 1, 0], [WIDTH - 1, HEIGHT - 1]], dtype="float32")
     M = cv.getPerspectiveTransform(src, dst)
     top_down_img = cv.warpPerspective(img, M, (WIDTH, HEIGHT))
 
     if DEBUG:
         cv.imshow("Top-Down Document", top_down_img)
-
         cv.waitKey(0)
         cv.destroyAllWindows()
 
@@ -104,40 +98,42 @@ def perspective_transform(img, doc_corners):
 
 
 def filter_binary(img):
-    print("Adaptive Thresholding")
+    # print("Adaptive Thresholding")
     # Adaptive Thresholding
     gray_img = cv.cvtColor(img, cv.COLOR_BGR2GRAY)
     binary_img = cv.adaptiveThreshold(gray_img, 255, cv.ADAPTIVE_THRESH_GAUSSIAN_C, cv.THRESH_BINARY, 11, 5)
 
-    cv.imshow("Final Document", binary_img)
+    if DEBUG:
+        cv.imshow("Final Document", binary_img)
+        cv.waitKey(0)
+        cv.destroyAllWindows()
 
-    cv.waitKey(0)
-    cv.destroyAllWindows()
-
-
-def scan(img, save=False):
-    doc_corners = find_document_corners(img)
-    top_down_img = perspective_transform(img, doc_corners)
-    binary_document = filter_binary(top_down_img)
-
-    if save:
-        output_filename = os.path.join(".", "results", "document.jpg")
-        cv.imwrite(output_filename, binary_document)
-
-    # if DEBUG:
-    #    cv.imshow("document", binary_document)
-    #    cv.waitKey()
+    return binary_img
 
 
 if __name__ == "__main__":
-    DEBUG = True
+    DEBUG = False
+    SAVE = True
 
-    test_folder = os.path.join(".", "imgs", "rotation")
+    test_folder = os.path.join(".", "imgs", "test")
     for image_name in os.listdir(test_folder):
         filename = os.path.join(test_folder, image_name)
         if not os.path.isfile(filename):
             continue
-        test_img = cv.imread(filename)
+        img = cv.imread(filename)
 
-        print(f"scanning document: {filename}")
-        scan(test_img)
+        print(f"\ntrying to scan document: {filename}")
+        doc_corners = find_document_corners(img)
+        if doc_corners is None:
+            print("FAILURE: unable to find a countour that can be approximated using 4 corner points!")
+            continue
+        top_down_img = perspective_transform(img, doc_corners)
+        binary_document = filter_binary(top_down_img)
+
+        if SAVE:
+            result_dir = os.path.join(".", "imgs", "results")
+            if not os.path.isdir(result_dir):
+                os.mkdir(result_dir)
+            output_filename = os.path.join(result_dir, image_name)
+            cv.imwrite(output_filename, binary_document)
+            print(f"SUCCESS: document has been saved to: {output_filename}")
